requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
APScheduler==3.10.4
asyncpg==0.29.0
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.24
botocore==1.35.24
certifi==2024.8.30
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
jmespath==1.0.1
lxml==5.3.0
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.11.1
